fix sf bug #525688 - parameter problems in sendfile testcases.  Should help on 64bit archs.

+++ testcases/kernel/syscalls/sendfile/sendfile02.c
@@ -47,6 +47,7 @@
 #include <errno.h>
 #include <fcntl.h>
 #include <sys/stat.h>
+#include <sys/sendfile.h>
 #include "usctest.h"
 #include "test.h"
 
@@ -99,7 +100,7 @@ main(int ac, char **av)
 	/*NOTREACHED*/
 }
 
-do_sendfile(int offset, int i)
+do_sendfile(off_t offset, int i)
 {
 	int in_fd, out_fd;
 	struct stat sb;
